backend/requirements.txt
# Core runtime dependencies (local/dev/prod-safe)
fastapi==0.110.1
uvicorn[standard]==0.34.2
motor==3.7.1
pymongo==4.10.1
pydantic==2.10.6
python-dotenv==1.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.5.0
PyJWT==2.11.0
python-multipart==0.0.20
httpx==0.28.1
email-validator==2.3.0
dnspython==2.8.0

# Excel parsing for catalog import
pandas==2.2.0
openpyxl==3.1.2

# Test/mock support
pytest==9.0.2
mongomock-motor==0.0.36
